Add access control

+++ OrganizadorReuniao/Models/User.cs
@@ -29,6 +29,24 @@ namespace OrganizadorReuniao.Models
             return database.executeQuery(cmd);
         }
 
+        public Result addUser(string email, string password, int unitId)
+        {
+            MySqlCommand cmd = new MySqlCommand("insert into lds_user (email, password, created_by) values (@email, @password, now())");
+            cmd.Parameters.AddWithValue("email", email);
+            cmd.Parameters.AddWithValue("password", common.hash(password));
+
+            Result result = database.executeQuery(cmd);
+            if (result.Success)
+            {
+                cmd = new MySqlCommand("insert into lds_user_unit (user_id, unit_id) values (@user_id, @unit_id)");
+                cmd.Parameters.AddWithValue("user_id", result.Id);
+                cmd.Parameters.AddWithValue("unit_id", unitId);
+
+                result = database.executeQuery(cmd);
+            }
+            return result;
+        }
+
         public Result updateUser(int id, string email, string password)
         {
             MySqlCommand cmd = new MySqlCommand("update lds_user set email = @email, password = @password where id = @id");
@@ -41,10 +59,36 @@ namespace OrganizadorReuniao.Models
 
         public Result deleteUser(int id)
         {
-            MySqlCommand cmd = new MySqlCommand("delete from lds_user where id = @id");
+            MySqlCommand cmd = new MySqlCommand("delete from lds_user_unit where user_id = @id");
             cmd.Parameters.AddWithValue("id", id);
+            Result result = database.executeQuery(cmd);
 
-            return database.executeQuery(cmd);
+            if (result.Success)
+            {
+                cmd = new MySqlCommand("delete from lds_user where id = @id");
+                cmd.Parameters.AddWithValue("id", id);
+
+                result = database.executeQuery(cmd);
+            }
+
+            return result;
+        }
+
+        public List<User> getWardUsers(int unitId)
+        {
+            List<User> users = new List<User>();
+            foreach (List<string> data in database.retrieveData("select us.id, us.email from lds_user us, lds_unit un, lds_user_unit uu " +
+                " where unit_id = @unit_id " +
+                "   and uu.unit_id = un.id " +
+                "   and uu.user_id = us.id " +
+                " order by us.email", unitId))
+            {
+                User user = new User();
+                user.Id = common.convertNumber(data[0]);
+                user.Email = data[1];
+                users.Add(user);
+            }
+            return users;
         }
 
         public User getUser(int id)
@@ -55,6 +99,9 @@ namespace OrganizadorReuniao.Models
                 user.Id = id;
                 user.Email = data[0];
                 user.Password = data[1];
+                Unit unit = new Unit().getUnit(id);
+                user.Unit = unit.Id;
+                user.UnitName = unit.Name;
             }
             return user;
         }
@@ -67,6 +114,20 @@ namespace OrganizadorReuniao.Models
                 user.Id = Convert.ToInt32(data[0]);
                 user.Email = email;
                 user.Password = password;
+                Unit unit = new Unit().getUnit(user.Id);
+                user.Unit = unit.Id;
+                user.UnitName = unit.Name;
+            }
+            return user;
+        }
+
+        public User getUser(string email)
+        {
+            User user = new User();
+            foreach (List<string> data in database.retrieveData("select id from lds_user where email = @email", email))
+            {
+                user.Id = Convert.ToInt32(data[0]);
+                user.Email = email;
                 Unit unit = new Unit().getUnit(user.Id);
                 user.Unit = unit.Id;
                 user.UnitName = unit.Name;
